# src/config/swagger.yaml
openapi: "3.0.0"
security:
  - passwordAuth: []
info:
  title: "DockStatAPI"
  version: "2.0.1"
  externalDocs:
    description: DockStat(API) Wiki
    url: https://outline.itsnik.de/s/dockstat
  license:
    name: BSD-3-Clause
    url: https://github.com/Its4Nik/dockstatapi/tree/main?tab=BSD-3-Clause-1-ov-file#readme
  contact:
    email: info@itsnik.de
  description: |-
    ![DockStat](https://github.com/Its4Nik/dockstatapi/blob/dev/.github/DockStat-dark.png?raw=true)

    # Pipelines

    [![Docker Image CI](https://img.shields.io/github/actions/workflow/status/Its4Nik/dockstatapi/build-image.yml?branch=main&label=Docker%20Image%20CI&style=for-the-badge&logo=docker)](https://github.com/Its4Nik/dockstatapi/actions/workflows/build-image.yml)
    [![Validation](https://img.shields.io/github/actions/workflow/status/Its4Nik/dockstatapi/validation.yml?branch=dev&label=Validation&style=for-the-badge&logo=checkmarx)](https://github.com/Its4Nik/dockstatapi/actions/workflows/validation.yml)

    # Feature List:

    - Swagger API Documentation
    - Database (Keeps data for 24 hours max)
    - Advanced authentication using hashes and salt
    - `http` API to configure the backend
    - Multi-arch docker builds (using buildx github action)
    - Advanced security through middlewares: rate-limiting and authentication
    - Multi Arch Docker builds through docker buildx
    - High Availability using single master and unlimited worker nodes!

    # 🔗 DockStatAPI v2 Documentation

    _⚠️ = Deprecation warning_

...

...

...

...

...

...
servers:
...
paths:
...
  /conf/addHost:
    put:
      tags:
        - "Configuration"
      summary: Adds a new host to the configuration and starts querying it
      operationId: addHost
      parameters:
        - name: name
          in: query
          required: true
          description: A name for the new host
        - name: url
          in: query
          required: true
          description: The target IP or dns entry
        - name: port
          in: query
          required: true
          description: The targets port on which Docker-Socket-Proxy runs
      responses:
        "200":
          description: Success - Host added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Host added successfully"
        "400":
          description: Error - Name, Port, and URL are required
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Name, Port, and URL are required"
        "401":
          description: Host already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Host already exists"
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /conf/removeHost:
    delete:
      tags:
        - "Configuration"
      summary: Removes an host from the config
      operationId: removeHost
      parameters:
        - name: hostName
          in: query
          required: true
          description: "The name of the to-be-removed-Host"
      responses:
        "200":
          description: Success - Host removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Host removed successfully"
        "401":
          description: Error - Host name is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Host name is required"
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "404":
          description: Error - Host not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Host not found"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
...
components:
  securitySchemes:
    passwordAuth:
      type: apiKey
      in: header
      name: x-password
      description: Password required for authentication
  schemas:
...
    402:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid password"
    403:
      type: object
      properties:
        status:
          type: string
          example: "denied"
        message:
          type: string
          example: "Password required"
    500:
      type: object
      properties:
        status:
          type: string
          example: "critical"
        message:
          type: string
          example: "Please see the server logs for more info"
    503:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Service unavailable. The high-availability lock is currently active. Please try again later."